Almost all of the new endpoints are downwards compatible with the old version. There are some  breaking changes in the certificates endpoint. Some values return a string with a comma separated list before and now return a proper JSON array.
The  goal is to have parity with the GUI and so new endpoints are added like subnets, domains, exposed email addresses and blacklisted assets. There is also a new endpoint you can use to check on how far your initial scan is. A code example on how to use it can be found in on_demand_scan.py.

Lastly, some endpoints have an extra option (full=True) that returns raw scan data. This can be useful for those of you who want to do data science on the results of the ShadowTrackr scans.

Want to know more? See https://shadowtrackr.com/docs/api